command system from jmod (ass) + module "improvement"
modules can now be added to a category from the super

this makes it so you have to manually instance the class, but it actually makes addons possible. if i wanted addon support the category.addModule method wouldve worked fine, but in addition to that you can specify the category you want from the category registry. it has a constant for each category instance. if youre using custom coded categories, store the instance of the category, register the category, then either: 1. reference that category in the super of your module. 2. call category.addModule(new x());. if you want to avoid having to manually create modules and add them one at a time, i didnt remove support for having a list of modules in the category super.

oh i also added CUtils from jmod. cool.